ai/requirements.txt
Flask==3.0.2
python-dotenv==1.0.1
pdfplumber==0.11.4
python-docx==1.1.2
pytest==8.2.1
